test(api): consolidate e2e access-token signing helpers
Replace repeated e2e JWT signing helpers with one support utility used across authenticated API test suites.\n\nRefs: #51
